dex_preopt_odex_install.mk: update for manifest_check.py changes.
Some of the uses-library entries in the manifest are propagated
transitively from dependencies (which are not necessarily uses-libraries
themselves, they may be e.g. static libraries). In Soong this is done at
deps mutator phase, but in makefiles it is done via dexpreopt.config
files (see dex_preopt_config_merger.py).

This patch updates makefiles for the changes in manifest_check.py:
  - option for dependency dexpreopt.configs is renamed
  - option for self dexpreopt.config is added (but not used yet)
  - 'Optional' field is added in makefile-generated dexpreopt.config
  - dex_preopt_config_merger.py is modified to add uses-libraries
    propagated via dexpreopt.config files

This patch does not enable generation of dexpreopt.config files for non
uses-library dependencies. This will be done as a follow-up.

Bug: 214255490
Test: lunch aosp_cf_x86_64_phone-userdebug && m && launch_cvd \
    && adb wait-for-device && adb root \
    && adb logcat | grep -E 'ClassLoaderContext [a-z ]+ mismatch'
    # empty output, no errors at boot
Change-Id: I6522319a8415f22f90fc96059a34675fb830e5cc