Add same_vndk_variants.timestamp to checkbuild
Before aosp/1851043, it was triggered by checkbuild because the
timestamp is attached to $LOCAL_BUILT_MODULE. With aosp/1851043, the
timestamp is not triggered by checkbuild and when a new VNDK lib is
added the check might get skipped in AOSP tree while it can break in the
downstream trees when actually used.

This change adds the same_vndk_variants.timestamp files to checkbuild
again so that CI catches the errors.

Bug: 233616647
Test: prebuilts/build-tools/linux-x86/bin/ninja -f \
  out/combined-aosp_cf_x86_64_phone.ninja -t query checkbuild \
  | grep same_vndk_variants.timestamp

Change-Id: I26043ec001ca2473fccfc19368d035374c663a21