Symlink farm creation tool from Bazel.
SymlinkTree is one of the action types Blaze returns in response to
`aquery` (action query) response. build-runfiles is tool copied from Bazel
to implement this action. It creates a symlink farm (i.e., a directory
hiearchy consisting of symlinks) from a given manifest. It is needed by
the mixed builds (in particular, a runtime symlink farm is needed to run
any Python application).

Bug: 232085015
Test: m USE_BAZEL_ANALYSIS=1 com.android.adbd
Change-Id: I9cfcb33cb7d0f63bd36ffd2b4101f53cfc6a42fc