Deduplicate multiple $(inherit-product) calls
The product configuration code already has support
for deduplicating common ancestors when you inherit
a file twice via a diamond pattern. However, if one
product inherits another twice, that will not be
deduplicated, leading to duplicate entries in list
variables.

Deduplicate these inherits to be consistent with the
duplicated ancestors.

The starlark product configuration already had this
deduplication, so it was creating a discrepency
between make/starlark.

Bug: 237019892
Test: ./build/bazel/ci/rbc_dashboard.py --quick on an internal-only product
Change-Id: I543a80746412ffcb9743203399413a0e707111e6