Export `host_required` property in module-info
Harness dependencies have to be listed in `host_required` property in some device-only module types, such as `android_test`. In order to setup testing enviroment correctly, this information needs to present in module_info.json and be exported to Bazel targets.

For example, `CtsLibcoreTestCases` has `cts-dalvik-host-test-runner` in `host_required` property, and it's required to set up testing environment in Atest Bazel mode.

Test: m out/target/product/coral/module-info.json
Bug: 220991354
Change-Id: I8d2ab0da04c9f4aff17d290999c69ffe0534729f