Tune R8 shrinking defaults for system_server
Disable R8 shrinking by default if there are additional SYSTEM_SERVER_JARS
that may be downstream from services.jar, or if the jar ordering is unrelaible.
The longer term solution will be to feed R8 the PRODUCT_SYSTEM_SERVER_JARS
explicitly during the shrinking phase, but for now this avoids potential issues
with such dowstream jars.

Bug: 240588226,235484969
Test: m + lunch various Arc/Wear/Auto targets
Change-Id: I1d824e1e61de65604123b65b25b192a8a32fd0c1