Suppress R8 warnings for support library annotations
As these annotations are stripped at compile-time, they are not
necessarily present in the generated dex library's classpath. As the
classpath is used to inform R8 about `libraryjars`, this can cause
reference failures during R8 when warnings are enabled.

For now, suppress the reference warnings, as we know these classes will
be stripped at compile-time. A proper solution will be to ensure the
appropriate annotation libs are fed as `-libraryjars` to R8.

Bug: 242088131
Test: m + `optimize { ignore_warnings: false }`
Change-Id: Ie92db2a9725ff520b311c6b66a50feae79463925