Revert "Add security patch level param to otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4d9d5e78-5aff-4592-97e0-0839e136ff9d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie15d9b17297d75ba9d65d6727ce330812c2476c2