Revert "Add security patch level param to otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/029a11ea-acd2-4035-b862-c55f196c3ac8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9a598b1a2a57e55e528ec778c85db281271ca93