Remove bootconfig cmdline arg from vendor_boot.img
This argument is being built into the GKI kernel, so devices no longer
need to pass it in from the vendor boot image.
If a device is using an older GKI kernel that isn't including this
argument, they will need to add the following for their device:
BOARD_KERNEL_CMDLINE += bootconfig
That will add the argument back to the vendor_boot.img.

Test: m && unpack_bootimg --boot_img=vendor_boot.img
Bug: 247548654
Change-Id: Ifeb68773af859ede48dfdc21cf08ae73e2d6c713