Revert "Use the toolchain from JDK 17"
Revert "Use the toolchain from JDK 17"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7fd3043a-26d5-4942-b8a1-6347075585f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I863b772120df8b546e35183ac659610f2b1f2f71