Revert "Use the toolchain from JDK 17"
Revert "Use the toolchain from JDK 17"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3be5987a-107e-4279-a8e1-aad4e83be2ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I666ca07e32ec7492278aa04098e935a64daf2b7d