Revert "Use the toolchain from JDK 17"
Revert "Use the toolchain from JDK 17"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/8f8ab801-03e0-41a1-bac7-68b50d5340ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I790383f24e8bea7784c6f3ea01ce5b9eecef3e34