Revert "Add more product defaults to mainline.mk"
Revert "Move PhotoTable to /product"

Revert "Remove duplicated product configuration"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2a95be5b-1675-486c-be77-44180f0dd748). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0b71c5a829dab6c1bff707c2ba31541e712f7f8