Revert "Use the toolchain from JDK 17"
Revert "Use the toolchain from JDK 17"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/5de5723f-c2ef-4dbe-b25f-82a4940a2532). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a38f0926eec00454e16a5a6ea7d6bc3f01619e3