Revert "Add more product defaults to mainline.mk"
Revert "Move PhotoTable to /product"

Revert "Remove duplicated product configuration"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2b99087e-0369-4d9f-931a-2be1bec13c35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iffb75a2aa3e138655f2a64627a69446a451146ef