Revert "Add more product defaults to mainline.mk"
Revert "Move PhotoTable to /product"

Revert "Remove duplicated product configuration"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bd9fef01-baea-454a-ac48-3b1f0453d770). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6eb4341489c12e74bf5f9a876761762b6d2b2197