Revert "Use the toolchain from JDK 17"
Revert "Use the toolchain from JDK 17"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7d0c6411-d5f5-49f8-8e46-dc257b97b82b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75153343a1d677da08d31ebd66555c85d109d6b4