Revert "Add more product defaults to mainline.mk"
Revert "Move PhotoTable to /product"

Revert "Remove duplicated product configuration"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/be470d75-d853-4553-9594-5618ff40d3ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I658bc669ff9ce575ac289950eb6d04e5b3ee2a6b