Install virtualmachine jar on /system_ext if product doesn't support AVF
In Android U we are planning to provide privileged @SystemApi's to
manage virtual machines.

These APIs will be provided by the android.system.virtualmachine jar,
which is provided by the com.android.virt APEX. The jar will also
contribute to the BCP.

However, since com.android.virt APEX is quite big, it will only be
installed on products that support AVF. For the rest of the products we
put the same android.system.virtualmachine jar on the /system_ext
partition.

Bug: 243512044
Test: build on product with AVF support
Test: build on product without AVF support
Change-Id: I994ef5bd7ea79ff1bdec846cf3b3ae0715c23263