Revert "check_target_files_vintf: Support flattened apexes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a0536446-0d9b-425c-8387-2bb69c8be732). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia57d6edb2403d6f93bd8ad0c874f76406773daaf