Revert "finalize branch: demarcate failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f832cb03-36f5-4f38-900a-21d6ac98de4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ffe49dafe87a716cbdd5ea0b40a63ef9a7a0eae