Revert "Fix typo"
Revert "Revert ^2 "Prevent unspecified values in soong_config_st..."

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9c434661-bd1b-4d04-a419-c8ac49cd7053). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e322f03962881d2e76edf6ba3951acbc14e8aa9