Revert "Fix typo"
Revert "Revert ^2 "Prevent unspecified values in soong_config_st..."

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/53da9908-86c7-45f2-85fe-b8f4bfcd8a1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I769b693bdc5ca27a1bd2a0b97e022fda18b2778c