Revert "Fix typo"
Revert "Revert ^2 "Prevent unspecified values in soong_config_st..."

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/33b1d9c9-2f88-491f-89ae-274db0ae07d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a6f69009f249f1e7292c776028933021fa7599e