Revert "Zip apex system server and standalone system server jars..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f157791f-e400-4343-80af-7162cb48ced0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I58dd2058e06b642d650b93ad40cb2e7b440ce49c