Allow empty modules.load file to be generated
Currently, if BOARD_*_KERNEL_MODULES_LOAD is set to empty string, we
default to load all modules listed in BOARD_*_KERNEL_MODULES. For
system_dlkm, we do not want to load any modules unless they are
dependency of vendor kernel modules. Therefore we need to create an
empty modules.load file for system_dlkm. Add 'false' as an out of band
value to indicate that an empty modules.load file should be generated.

Bug: 149866755
Test: th
Change-Id: Ieaaeb2f9a824e97671acff465bbd4da83dfd9aa0