Revert "IAllocator-V2"
Revert "IMapper 5 - the Stable C approach"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffde8f40-05ad-47dd-aa39-6e56ca7ce7e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3727b65cd9c2f63e1d63d25c76e43d32807d53c1