Revert "IAllocator-V2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/598e2575-33ec-42bd-83ab-75c6c9a82d59). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iccd43a4d165216d4de276eab60f3c47233aff05d