Revert "Place unstripped pvmfw image to the symbols directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d90823a0-c044-477e-ac7c-f702d4e0a2f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c02a9594be64b70d6bda9a8cdead77d949beccf