Revert "Handle non-module fsv_meta files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2eb28559-b5a9-4663-b1d1-f008d0dcc569). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iadf8d89230a4cb4d43bdeebd2e77da6e4f5fff4e