Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36447f78-5604-41f0-9c7e-e9fd60b888fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c05c6fb2b4105baeef08caf806d1150aa634c40