Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae80bd79-5c4d-485c-aae1-bf9eb2f753ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f2ef409c41b030b006bb641b699784dc182bee0