Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff8d884e-f574-45ff-8ffc-a7a3c13d5c29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I92605e0dff7af86396f9b2ec467c9702b8ac7791