Revert "IAllocator-V2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd8b0b85-865c-4354-8a6d-ca722490f9e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b8811359c35fe787ba31297c4bd7fce8eb5c294