Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/996da495-1806-4e29-b1a8-849e06efbff2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6a0c4233cdd4dac260466e82abb3a3696c9d7a5e