Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b61a0056-2e83-4ef7-bb6a-108ba86ea8b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefb56df58a3fb39d0e7d9ae143789d8f2623eaf4