Revert "Replace the HIDL cas service from default vendor images"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2632b4ee-904c-4db5-80ef-cf6398190d4c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie21fcda45707d21da9159e5e1895c9c0db0128f0