Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60465e80-48b4-4d3a-a5a2-69f967349e40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic2260a36eb4aff793f35b90d315aacac9a10912e