Revert "Fix non-AB OTA signing issue"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/619a73ae-db28-419c-9cd9-afea34019580). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c56629f773f1bb7a9a84a6a909c434ca2fe16b9