Revert "Package blkid_static in otatools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01a9d59f-1ae6-404b-b4ed-1f473cae5c3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I644ec16c488c2e0b5e5ddc00ecffce38404689af