Revert "Don't disable dexpreopt for an unbundled build."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a0e134c-8bdd-48bf-a793-26e23008aaa2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If35c4ccca16c61f842b0bd3649f187238b844bab