Revert "Change function ZipDelete to use Python module zipinfo instead of command 'zip'."
This reverts commit 8895f9eceac4ad568bcca724c4dcf8faf5b865f4.

Reason for revert: New ZipDelete implementation is extremely slow. Running zip -d on a 4GB target_files.zip took 8 seconds, it takes ZipDelete() 500 seconds to achieve the same thing. This is because ZipDelete() works by copying every entry to a new zip file, decompressing and recompressing everything in the process.

Change-Id: I0548255bc29380303314763f6d81e74bf3dbb76e