Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9f510a1-e5f6-435a-a01a-3b07c3936d6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9ce3f3e0f58c88347e2a7ad0fb0a9ed56cd4e3b