Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bec75cd2-fdb0-4e18-a958-b8d6a90d6be5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31c99e2b8cd179529bac94499485d9c87b1c598c