Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e05dacb-2c0d-44bd-a0f6-99726ab1a085). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I735ba45be479f5402818dd0629cb3f1a3a0c694a