Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/762fa49f-4f2a-4075-b57f-9f89c7b4dfac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I35f34b4b5c2c58bd72b5e9251a71e12e1fc815e8