Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7125fcc7-55f3-444f-9c61-9127f2a01b9a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I756b97c459c69c19f5a9b2fad6a02acb32a8ba45