Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95f7d62a-3dc2-45d3-a635-4accc30d3cb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e1fa9ef82d3f3303dc076badff868a48aec407a