Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2039385-a760-41fb-a166-8d0865ab4ffb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbc54d2cf421b6cb7bf7f4f84b1dc3abc5004472