Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4377a8a-1544-4836-aaf7-06f364a82e42). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1de93a44ce5f478b346de7ad686896470866850d