Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/534d4506-ed8b-4009-a8f2-a9a4d9e48236). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60ac7df587097f3491e0e9f02e11096f4df6c804