Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b15a89e8-36cb-46d7-88ad-e7861032d348). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5339cabb19b774d1aa168a7e4731020a6a0b757b