Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f17bfce-a7c1-46e2-8120-289b1f81550b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If04d0af5d6efe0e0a8bfce981634f2eecd57132e