Revert "Disable dexpreopt for aosp_riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63b6932f-5abf-4dd6-8df6-f11f90596dbc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2db65f03726eceabe2ede4cb9a16a5e63de48b70