aosp_arm64: Build pvmfw.img during GSI builds
As the interfaces between host code contained in the system.img and
guest code contained in the pvmfw.img are not stable, flashing GSI on a
device supporting the Android Virtualization Framework (AVF) may break
its support for booting virtual machines. Instead, build the pvmfw.img
for GSI releases so that it can also be flashed along system.img.

Bug: 272443823
Test: TH
Change-Id: Ie109a4364d1f6d00f47ba795643efdc30ea777cf