Revert "Generate SBOM of the target product in file sbom.spdx in..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b65d203-d364-4ade-aabb-1330fe45236a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c43095894ad916ac08282cbe045d6a21ac3d1b6