Revert "Generate SBOM of the target product in file sbom.spdx in..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c58e49bd-e353-4b16-9b51-e2465d2a0349). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia25618f4fc03d11b2b4432fa036a38fec28c7c3d