Revert "Generate SBOM of the target product in file sbom.spdx in..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63eb27f5-51a4-4789-b499-63a2cf36aec8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icaaf5f90e0c1cf2343bd7953dbcfddf67844b253