Revert "Generate SBOM of the target product in file sbom.spdx in..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc62e915-6b58-472c-8e73-3edd0bf9b34f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I388ad6f8a0e4fd2c2129e5e2514aeaa061f70439