Revert "Don't install jni libs for uninstallable modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c752f5d1-61f1-474e-ae53-448cdfe18d93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie5b5df7d35c3730802edd619219ca6f44d68d0d2