Revert "Don't install jni libs for uninstallable modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/127681c2-561d-4780-8326-d3ff74371f93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe8cdab4999246c0fa8ecd25598a4f139dfa7f73