Revert "Don't install jni libs for uninstallable modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8237886-4795-419f-a307-f78fcfd1f0d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15dc9cc7f8c2ba69d31a0a90ebaeb4d1f818be34