Revert "Don't install jni libs for uninstallable modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df3e6f29-eea0-4a69-808b-58df66127971). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie27fbbd94ae15b640074a1acb1b8dc24e67ce2bb