Revert "Don't install jni libs for uninstallable modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/96fab15e-ed09-4a52-ab1b-fc3c039a7737). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39c5a5a0adbc5736efd2aae923d95ca511d5b22b