Revert "remove reference to old VTS config"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91c8739e-a2e0-41d2-a4d3-b9eb1c89fbd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d5b319dfe48f95c5b899a4377ea3e4c22f1dae9