Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ce1b2ba-cd52-45be-861e-88a2801effde). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I13df28b9cb8d015fa91cf24913856e58d1ce8e10