Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c2c3328c-e477-4df8-8988-62d1c6e35b21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3a12f1202e46476666c86eea7e2c35e65f36433