Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e268ec7b-7b30-4d75-87ac-423c795e6912). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I231e7d9baccb32c7811afd973f67d677cdfe4e7f