Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/624a39e0-0e95-40f1-88b1-e3d79f182054). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia60a8a007e75f23ea834dcfbe893cc75f117ef33