Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1678d03d-9737-432b-9162-f9e7530a5bce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5493a2e725928a08961ae0625ce50e6bbfb36ffa