Revert "add PRODUCT_SOURCE_ROOT_DIRS product variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d717180-c435-4425-9cb4-ca0c95a23046). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I819c02f24b8621c7a44086bb6138a5b39239f1f8