Revert "Copy SBOM to dist directory when running "m dist"."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48542528-78a8-4253-a48f-c3e01d052668). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I327ef9c55ce10fc1c4c72d6d282176152f363b5c