Revert "Invoke delta_generator directly"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a49b54d3-fa9e-4285-a6c3-047c1906e38e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5135c88bcf7347f6765aba96c64c8a88f3317a8