Revert "Invoke delta_generator directly"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d4ede63-d932-46c4-82b9-91cbbbed5a31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia702837b693813107dcd2750690c52b170269023