Fix expansion of single-value product variables
If a single-value product variable is set in multiple inherited product
makefiles it currently inherits the value from the makefile that comes
first in lexical order as opposed to the value from the first inherited
makefile that sets it. Fix this so that a single-value product variable
inherits its value from the first inherited makefile that sets it by
making sure the variable expansion in _expand-inherited-values happens
in order of makefile inheritance as opposed to lexical order.

Bug: 277793604
Test: Output of lunch aosp_cheetah-userdebug && \
  build/soong/soong_ui.bash --dumpvars-mode \
  --vars="$(get_build_var _product_var_list)" is identical before and
  after patch
Test: Patched aosp_cheetah.mk to inherit foo.mk followed by bar.mk
  setting PRODUCT_NAME_FOR_ATTESTATION to foo and bar respectively and
  lunch aosp_cheetah-userdebug && \
  get_build_var PRODUCT_NAME_FOR_ATTESTATION returns bar before this
  patch and foo after this patch

Change-Id: I1b936f189d7836140b0c1d0528f34633d6160bfe
Signed-off-by: Rashid Zaman <rashidz@meta.com>