Revert "Invoke delta_generator directly"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a3b80698-b90d-4d58-9d6e-5d5209818281). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bde6dfda0398ec13bbfd9ff7baf317d1d92fd50