Revert "Remove lunch target from appcompat log"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/552776dc-8ecb-45ac-937a-406f68eccab2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a4f687220607298b322d244cd158f6d2c151aaa