Revert "Update PLATFORM_VERSION to VP1A"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7cf48bc3-b822-4ce8-aeb3-4687efd1c759). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c90428ee9ab4d9c60d5fe160218ebdd8d21729a