Revert "Update PLATFORM_VERSION to VP1A"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27b7aad6-3582-4b56-a7b9-db70366523e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1171ccb3cdee524acf073e97493c9133056dd50