Revert "Update PLATFORM_VERSION to VP1A"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ace45e4-e070-4348-95f0-7dcdc21f51e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53a35613cd8f0d1318c6c54a5828f36d86e267a9