Revert "Remove emulator dependencies on non emulator targets (2)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4849e447-de9c-4ba3-b19e-0a6579bc84ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2638deea88ee4a7da2c514d11f933e2f32c56b33