Revert "(reland) Remove emulator dependencies on non emulator ta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f110142d-3213-44ad-b3c5-ffb348930e0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5323edc1f423ede3e1e18210f3389a5d4eb74d1c