Revert "(reland) Remove emulator dependencies on non emulator ta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7904842e-7c1a-4133-a09f-a36d833f1ad0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If62ec968150458eac4737f05e9d6d2f3fd33821b