Revert "(reland) Remove emulator dependencies on non emulator ta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4174a2df-58d0-4853-8413-3d3259990c97). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54f6b932c4eb482cfc0038d97d4f5df5a009bcc3