Revert "(reland) Remove emulator dependencies on non emulator ta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f81759e1-5a36-4231-9649-1de6a0f3dcce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If4f7f34c6995853927607ff2b00343480e84234d