Revert "Parallelize generation of OTA package and target_file pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e1e087d-7d42-4939-8edb-d31c5f954e24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6624e78602fd7f14f9fe58dae4df9f640fdfcc27