Revert "Parallelize generation of OTA package and target_file pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d612d043-e4cf-4250-86a9-6e2b29f1a3ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c9af7a2a788930c59714af7c88706beb64b3a47