aconfig: Add first iteration of cpp codegen to aconfig
The general idea to reuse java codegen's very neat tiny template idea.
For generated cpp code, it is in the form of a collection of classes
inside a namespace. The reason we choose a collection of classes rather
than a collection of static functions is because gmock test technology
only supports mocking virtual method.

BUG: b/279483801
TETS: atest aconfig.test

Change-Id: I9ba00667437ff7c3e147ff2828171fc95528bebf