Revert "parallelize add_partition_calls"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4acc9ae9-e73f-4a1d-a80f-4c75f1e7c97e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I14b761ad675c8c1a6024431f2c20d321116aa5c1