Revert "parallelize add_partition_calls"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c503ff6c-032b-42bf-a43d-ef0229195dd9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38b6b01656711231f4349b2c21a464f2911d2037