Revert "Adding vendor_kernel_boot to fasboot-info"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a15904d-8154-4db2-9713-5c3d5a5ff2fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb36b065e8df76f118f1752492d418f96e56a4cc