Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e457d2c-c4e1-4599-92ed-123d62392a8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6208edc890607492abf1ea0ad6a2b6a5237432e