Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf83aa8b-0a17-4368-9949-938e4738a02b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6da44301a18f4033e28848080b7bb82f6f0232e6