Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9591e083-30b6-4738-ba8d-44b9591fb651). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2acc9178069b1eb71f17401a303c4cbe66f81e42