Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ffe4fa1-031e-4946-bdad-5eb06e8928e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2982d169a584212d4a4ae9268ceb0fef7fd9c2e7