Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e5b715d-da14-40ee-b366-a518d92b3262). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26582dde7ef9940098e6a8bdf38c6afe7e431eb3