Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24fa123c-3d59-4a34-b031-b95d3efaf70d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d243978ebb3ad8358aaa3d96411f0c4c540db47