Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe795b80-4f6c-4ce7-87ed-6cc7826462fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9dbbe74bce55a696bb91ca6e7510958ba0d824e2