Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1cfe3fc-36e9-4d94-ac13-c5f452dbb712). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a4be92414ebf2fbcc37646682de91ca61ef2318