Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2324ddbe-93bb-4f54-a6a9-27d8b72a32cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcb36ac736c7089e2de9c85ba8936aa77e34b8c3