Revert "Update LLNDK list"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ede4cb7c-8afd-4386-96ac-738219431cdb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d034d383ea9c5568c0cc3f4bdd0ec43d20ff2b9