Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/625e3f8d-13ce-4bc7-a7c5-991f518fe114). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24f275a97dd46a462c981a5d04672f2ebcbf73c5