Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0a096cc-5d6a-4648-a62c-a26102c31283). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf81b5d58824b573f0951a8fdf0ccd045281800f