Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa254277-2a4a-4e07-b6aa-f9051e30f64e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I082ef6b624b9fcfc1ea8c49fbf8c9a123b2985eb