Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e347ad4f-c007-48ec-8652-922c7a519413). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3cce1f5b32e102e1509bf393381b8ae5551d9a38