Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c32a908-2227-458a-b8ab-f6cd9e2c0d12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5f917e20cb2ea0c0e6d7bc2b018faf52f1bf122e