Revert "[MTE] [gnu make] Add memtag_globals as a sanitizer target."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da9517c5-93e5-47ac-9ee9-0611a804b0ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I758a5a4068ad620648bcb25b9ab45b5b2115afd6