Revert "Enable CFI for the whole Bluetooth repo and not just system"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f13013f-42ea-4321-9c1c-fac009c489c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d9b0f6c487e55ef695a6f22365ce65012bb2316