aconfig: Add namespace constant in generated java code
(Package name + flag name) is the unique id of a flag, but it could
happen that two flags have the same package name and flag name, but
different namespaces and are built into different libraries. There will
be two cases

* These two libraries are used in the same app/process, the first one
  in the path will be loaded. So the second one will miss without error.
* These two libraries are used in different app/processes. It won't
  cause issue, but it will be confusing.

This change adds a namespace constant value in the generated java code.
It won't solve the issues above, but it provides a way to check the
unique id of a flag, which will be helpful in test.

Test: atest aconfig.test
But: 279483816
Change-Id: I1b10e2f4083df3e5e50e1a049b4c8683e200ef41