Revert "Replace libc++.so.1 with libc++.so"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0734c747-6232-435c-b5d0-be1d9b0b6abb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4cfb5c44920e54f7c984e4bdbce88099bf6ac0d8