Revert "Replace libc++.so.1 with libc++.so"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b8ab218-968d-4dca-9beb-38e3502a9fa1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56fff5c564a7019804d90ac4184beea4fef51d12