Revert "Remove dmtracedump from build/make"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7945fb5d-2336-42e7-a2be-3284652dc9be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf17eb893ee0da3c77941158a931ac35020fd035