Revert "Add more product defaults to mainline.mk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b48c27b-82e0-42ca-9a53-54c38745fb4e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4dae7ab0f1979e883f573408920bcf3d8b4fedf1