Revert "Add more product defaults to mainline.mk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/043ac244-8931-4ac1-8b97-5e7d5ae52651). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If8995dc683d27e561cf2753813a1361bead4b6c1