aconfig: limit max number of bug fields to one
Some consumers of `aconfig dump` do not support multiple bugs. To
prevent this from becoming a problem, limit the number of times the bug
field in the flag_declaration and parsed_flagp proto messages may
appear to exactly zero or one.

This change could have been implemented as a change of `repeated` to
`optional` in the proto definition. However, the chosen approach, with a
runtime check, is easier to revert if we want to support multiple bugs
in the future.

Bug: 293156797
Test: atest aconfig.test aconfig.test.java aconfig.test.cpp
Change-Id: Ib87dac68b392986a8daa64e56cd85477c92fbe83