Revert "Support transitive resource packages from Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3380f0d9-b439-4c2d-b2a9-103c4d2db42c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0f7e537785e24c597dbed5000f7d0dfca0f8379